# skills-matrix-be/src/config/swagger.yaml
...
openapi: 3.0.0
info:
  title: Swagger — Skills Manager Suite
  version: 0.0.1
  description: An API for effective skills management in project teams
  contact:
    name: Jaona Ramahaleo
    email: jaona.ramahaleo@gmail.com
  license:
    name: MIT
    url: http://tldrlegal.com/license/mit-license
servers:
- url: http://localhost:3000/v1
  description: Local Development Server
paths:
...
  /projects/{project_id}:
    get:
      tags:
      - Projects
      parameters:
      - name: project_id
        description: The ID of the project
        schema:
          minimum: 1
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: OK
        "404":
          description: Project not found
        "500":
          description: An error occurred
      operationId: getProject
      summary: Retrieve details about one project
    put:
      requestBody:
        description: Project to update to the database
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Project'
        required: true
      tags:
      - Projects
      parameters:
      - name: project_id
        description: The ID of the project
        schema:
          minimum: 1
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: OK
        "404":
          description: Project not found
        "500":
          description: An error occurred
      operationId: updateProject
      summary: Update a project
    delete:
      tags:
      - Projects
      parameters:
      - name: project_id
        description: The ID of the project
        schema:
          minimum: 1
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: Project deleted
        "404":
          description: Project not found
        "500":
          description: An error occurred
      operationId: deleteProject
      summary: Delete a project
    parameters:
    - name: project_id
      description: The ID of the project.
      schema:
        type: integer
      in: path
      required: true
  /projects/{project_id}/competencies:
    get:
      tags:
      - Project Competencies
      parameters:
      - name: project_id
        description: The ID of the project
        schema:
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getProjectCompetencies
      summary: List project competencies
    parameters:
    - name: project_id
      description: The ID of the project.
      schema:
        type: integer
      in: path
      required: true
  /projects/{project_id}/positions:
    get:
      tags:
      - Project Positions
      parameters:
      - name: project_id
        description: The ID of the project
        schema:
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getProjectPositions
      summary: List project positions
    post:
      tags:
      - Project Positions
      parameters:
      - name: project_id
        schema:
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: The request was fulfilled.
      operationId: createProjectPosition
      summary: Create project position
    delete:
      tags:
      - Project Positions
      parameters:
      - name: project_id
        description: The ID of the project.
        schema:
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: The request was fulfilled.
      operationId: deleteProjectPositions
      summary: Delete all project positions
    parameters:
    - name: project_id
      description: The ID of the project.
      schema:
        type: integer
      in: path
      required: true
  /positions/{position_id}:
    get:
      tags:
      - Positions
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getPosition
      summary: Retrieve details about a position
    put:
      tags:
      - Positions
      responses:
        "200":
          description: The request was fulfilled.
      operationId: updatePosition
      summary: Update details about a position
    delete:
      tags:
      - Positions
      responses:
        "200":
          description: The request was fulfilled.
      operationId: deletePosition
      summary: Delete a position
    parameters:
    - name: position_id
      description: The ID of the position.
      schema:
        type: integer
      in: path
      required: true
  /positions/{position_id}/competency_groups/:
    get:
      tags:
      - Position Competency Groups
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getPositionCompetencyGroups
      summary: List competency groups for a given position
    post:
      tags:
      - Position Competency Groups
      responses:
        "200":
          description: The request was fulfilled.
      operationId: createPositionCompetencyGroup
      summary: Create competency group for a given position
    parameters:
    - name: position_id
      description: The ID of the position.
      schema:
        type: integer
      in: path
      required: true
...
  /positions/{position_id}/requirements/:
    get:
      tags:
      - Position Requirements
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getPositionRequirements
      summary: List competency requirements for a given position
    post:
      tags:
      - Position Requirements
      responses:
        "200":
          description: The request was fulfilled.
      operationId: createPositionRequirement
      summary: Create competency requirement for a given position
    delete:
      tags:
      - Position Requirements
      responses:
        "200":
          description: The request was fulfilled.
      operationId: deletePositionRequirements
      summary: Delete all competency requirements for a given position
    parameters:
    - name: position_id
      description: The ID of the position.
      in: path
      required: true
  /requirements/{requirement_id}:
    get:
      tags:
      - Position Requirements
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getPositionRequirement
      summary: Retrieve details about a competeny requirement for a given position
    put:
      tags:
      - Position Requirements
      responses:
        "200":
          description: The request was fulfilled.
      operationId: updatePositionRequirement
      summary: Update a competency requirement for a given position
    delete:
      tags:
      - Position Requirements
      responses:
        "200":
          description: The request was fulfilled.
      operationId: deletePositionRequirement
      summary: Delete a competency requirement for a given position
    parameters:
    - name: requirement_id
      description: The ID of the requirement.
      schema:
        type: integer
      in: path
      required: true
  /projects/{project_id}/teams:
    get:
      tags:
      - Project Teams
      parameters:
      - name: project_id
        description: The ID of the project.
        schema:
          type: integer
        in: path
        required: true
      responses:
        "200":
          description: The request was fulfilled.
      operationId: getProjectTeams
      summary: List project teams
    post:
      tags:
      - Project Teams
      responses:
        "200":
          description: The request was fulfilled.
      operationId: createProjectTeam
      summary: Create project team
    delete:
      tags:
      - Project Teams
      responses:
        "200":
          description: The request was fulfilled.
      operationId: deleteProjectTeams
      summary: Delete all project teams
    parameters:
    - name: project_id
      description: The ID of the project.
      schema:
        type: integer
      in: path
      required: true
...
components:
  schemas:
...
    Project:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        project_otp_code:
          type: string
        start_date:
          format: date-time
          type: string
        end_date:
          format: date-time
          type: string
...